coach card, functional. cod refactored
